move Protector config related classes to own directory
